backend/requirements.txt
channels==4.2.2
Django==5.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
pillow==11.2.1
PyMySQL==1.1.1
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
requests==2.32.3
whitenoise==6.9.0
elevenlabs
gradio_client
openapi
google-genai
websocket-client